requirements.txt
google-cloud-vision==2.4.2